pyproject.toml
[project]
name = "mvp"
version = "0.1.0"
description = "Singapore Supply Chain Resilience Management System – MVP backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115,<1",
    "langgraph>=1.1.3",
    "python-dotenv>=1.0.1,<2",
    "uvicorn[standard]>=0.34,<1",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "basic"
venvPath = "."
venv = ".venv"

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "ruff>=0.15.8",
]
